升级了jsoncpp 0.5.0版本到1.8.4版本

+++ chilli/EeventReport/EventReportModule.h
@@ -59,7 +59,7 @@ private:
 	void ListenTCP(uint32_t port);
 	bool listenWS(int port);
 
-	std::map<uint32_t, EPConnectionPtr>m_Connections;
+	std::map<uint64_t, EPConnectionPtr>m_Connections;
 
 };
 }
